another conoutf fix

+++ src/console.cpp
@@ -97,7 +97,7 @@ void bindkey(char *key, char *action)
         strcpy_s(keyms[i].action, action);
         return;
     };
-    conoutf("unknown key \"%s\"", (int)key);   
+    conoutf("unknown key \"%s\"", key);   
 };
 
 COMMANDN(bind, bindkey, ARG_2STR);
